Move `to_internal_value` to HitasModelSerializer
