+++ backend/hitas/views/property_manager.py
@@ -1,6 +1,3 @@
-from uuid import UUID
-
-from hitas.exceptions import HitasModelNotFound
 from hitas.models import PropertyManager
 from hitas.views.utils import AddressSerializer, HitasModelSerializer, HitasModelViewSet
 
@@ -16,15 +13,6 @@ class PropertyManagerSerializer(HitasModelSerializer):
             "name",
             "email",
         ]
-
-    def to_internal_value(self, data):
-        """If referenced in another serializer by id, return an object"""
-        if data.get("id", None) is not None:
-            try:
-                return self.Meta.model.objects.get(uuid=UUID(hex=str(data.get("id", None))))
-            except (self.Meta.model.DoesNotExist, ValueError):
-                raise HitasModelNotFound(model=self.Meta.model)
-        return super().to_internal_value(data)
 
 
 class PropertyManagerViewSet(HitasModelViewSet):
