+++ backend/hitas/views/utils/serializers.py
@@ -1,5 +1,8 @@
+from uuid import UUID
+
 from rest_framework import serializers
 
+from hitas.exceptions import HitasModelNotFound
 from hitas.models._base import ExternalHitasModel
 from hitas.views.utils import PostalCodeField
 
@@ -9,6 +12,18 @@ class HitasModelSerializer(serializers.ModelSerializer):
 
     def get_id(self, instance: ExternalHitasModel) -> str:
         return instance.uuid.hex
+
+    def to_internal_value(self, data):
+        """If referenced in another serializer by id, return an object"""
+        if data.get("id", None) is not None:
+            try:
+                return self.Meta.model.objects.get(uuid=UUID(hex=str(data.get("id", None))))
+            except (self.Meta.model.DoesNotExist, ValueError):
+                raise HitasModelNotFound(model=self.Meta.model)
+        return super().to_internal_value(data)
+
+    class Meta:
+        model = None
 
 
 class AddressSerializer(serializers.Serializer):
